webapp/ansible/roles/webapp/files/app/requirements.txt
gunicorn==19.9.0
Flask==1.0.2
mysqlclient==1.4.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
flask_paginate==0.5.2
Pillow